move AddTask to cxx for Root6 compatibility
